+++ src/index.ts
@@ -12,6 +12,7 @@ export * from './components/SPopover';
 export * from './components/SSwitch';
 export * from './components/SFilter';
 export * from './components/SRadioGroup';
+export * from './components/SLabel';
 
 export { default as SSidebar } from './components/SSidebar.vue';
 export { default as SSidebarLogo } from './components/SSidebarLogo.vue';